@@ -2,7 +2,7 @@
   "name": "panamia",
   "version": "0.1.0",
   "private": true,
-  "type": "commonjs",
+  "type": "module",
   "scripts": {
     "dev": "env -u NODE_ENV node server.js",
     "dev:http": "env -u NODE_ENV next dev",
@@ -37,30 +37,30 @@
     "@stripe/react-stripe-js": "^2.3.2",
     "@stripe/stripe-js": "^2.1.11",
     "@tabler/icons": "^1.107.0",
-    "@tanstack/react-query": "^5.14.0",
-    "axios": "^1.2.1",
+    "@tanstack/react-query": "^5.90.12",
+    "axios": "^1.13.2",
     "bcrypt": "^5.1.0",
     "busboy": "^1.6.0",
     "class-variance-authority": "^0.7.1",
     "clsx": "^2.1.1",
     "date-fns": "^4.1.0",
-    "dotenv": "^16.3.1",
-    "formik": "^2.2.9",
+    "dotenv": "^16.6.1",
+    "formik": "^2.4.9",
     "framer-motion": "^7.6.12",
-    "lucide-react": "^0.554.0",
-    "mongoose": "^8.0.3",
+    "lucide-react": "^0.557.0",
+    "mongoose": "^8.20.2",
     "nanoid": "^5.1.6",
-    "next": "^15.5.6",
+    "next": "^16.0.8",
     "next-auth": "^5.0.0-beta.30",
     "next-sitemap": "^4.2.3",
     "next-themes": "^0.4.6",
-    "nodemailer": "^6.10.1",
-    "pigeon-maps": "^0.21.3",
+    "nodemailer": "^7.0.11",
+    "pigeon-maps": "^0.21.6",
     "pusher": "^5.2.0",
     "pusher-js": "^8.4.0",
-    "react": "^19.2.0",
-    "react-day-picker": "^9.11.3",
-    "react-dom": "^19.2.0",
+    "react": "^19.2.1",
+    "react-day-picker": "^9.12.0",
+    "react-dom": "^19.2.1",
     "react-dropzone-uploader": "^2.11.0",
     "react-ga": "^3.3.1",
     "react-hook-form": "^7.67.0",
@@ -74,13 +74,13 @@
     "stripe": "^14.9.0",
     "tailwind-merge": "^3.4.0",
     "xlsx": "https://cdn.sheetjs.com/xlsx-0.20.1/xlsx-0.20.1.tgz",
-    "yarn": "^1.22.19",
+    "yarn": "^1.22.22",
     "zod": "^4.1.13"
   },
   "devDependencies": {
     "@playwright/test": "^1.57.0",
     "@types/bcrypt": "^5.0.0",
-    "@types/busboy": "^1.5.3",
+    "@types/busboy": "^1.5.4",
     "@types/node": "18.11.5",
     "@types/pusher-js": "^5.1.0",
     "@types/react": "^19.2.7",
@@ -88,15 +88,15 @@
     "@types/stripe": "^8.0.417",
     "autoprefixer": "^10.4.22",
     "eslint": "8.26.0",
-    "eslint-config-next": "^15.5.6",
+    "eslint-config-next": "^16.0.8",
     "husky": "^9.1.7",
     "lint-staged": "^16.2.7",
     "mongodb-memory-server": "^9.5.0",
     "postcss": "^8.5.6",
-    "prettier": "^3.6.2",
-    "prettier-plugin-tailwindcss": "^0.7.1",
-    "tailwindcss": "^3.4.17",
-    "tailwindcss-animate": "^1.0.6",
-    "typescript": "^5.3.3"
+    "prettier": "^3.7.4",
+    "prettier-plugin-tailwindcss": "^0.7.2",
+    "tailwindcss": "^3.4.18",
+    "tailwindcss-animate": "^1.0.7",
+    "typescript": "^5.9.3"
   }
 }
