@@ -1,5 +1,5 @@
 /** @type {import('next-sitemap').IConfig} */
-module.exports = {
+export default {
   siteUrl:
     process.env.NEXT_PUBLIC_HOST_URL ||
     process.env.SITE_URL ||
